+++ src/schemapack/__init__.py
@@ -21,6 +21,7 @@ from ._isolate import isolate, isolate_class, isolate_resource
 from ._load import load_datapack, load_schemapack
 from ._main import load_and_validate
 from ._normalize import denormalize
+from .validation import SchemaPackValidator
 
 __all__ = [
     "load_and_validate",
@@ -30,6 +31,7 @@ __all__ = [
     "isolate_class",
     "isolate_resource",
     "denormalize",
+    "SchemaPackValidator",
 ]
 
 __version__ = version(__package__)
