+++ src/schemapack/validation/__init__.py
@@ -16,5 +16,16 @@
 
 """Logic for validating a datapack against a schemapack."""
 
-# shortcuts:
-from schemapack.validation.main import SchemaPackValidator  # noqa: F401
+from ._base import (
+    ClassValidationPlugin,
+    GlobalValidationPlugin,
+    ResourceValidationPlugin,
+)
+from ._main import SchemaPackValidator
+
+__all__ = [
+    "SchemaPackValidator",
+    "GlobalValidationPlugin",
+    "ClassValidationPlugin",
+    "ResourceValidationPlugin",
+]
