+++ src/schemapack/validation/_base.py
