+++ src/schemapack/validation/_plugins/one_to_many_overlap.py
@@ -22,7 +22,7 @@ from collections.abc import Mapping
 from schemapack.exceptions import ValidationPluginError
 from schemapack.spec.datapack import DataPack, Resource, ResourceId
 from schemapack.spec.schemapack import ClassDefinition
-from schemapack.validation.base import ClassValidationPlugin
+from schemapack.validation._base import ClassValidationPlugin
 
 
 class TargetOverlapValidationPlugin(ClassValidationPlugin):
